Fixed compilation warning.
git-svn-id: http://svn.tilp.info/repos/tiemu@2770 4f589446-7ac1-0310-82c7-a055b08196ef

+++ tiemu/trunk/src/core/dbg/bkpts.c
@@ -567,8 +567,8 @@ int ti68k_bkpt_get_pgmentry_offset(unsigned int id, uint16_t *handle, uint16_t *
 		return -1;
 	
 	data = GPOINTER_TO_INT(g_list_nth(bkpts.pgmentry, id)->data);
-	*handle = data >> 16;
-	*offset = data & 0xffff;
+	*handle = (uint16_t)(data >> 16);
+	*offset = (uint16_t)(data & 0xffff);
 	return 0;
 }
